Refactoring baseline codebase
* Tried to remove tight couplings

* Introduced makefile

* Added gitignore

TODO: Lot of refactoring still possible
