+++ .gitignore
@@ -1,0 +1,3 @@
+.vscode
+bin/
+aixi
