+++ common/types.hpp
@@ -1,0 +1,23 @@
+#include <map>
+#include <vector>
+
+// symbols that can be predicted
+typedef bool symbol_t;
+
+// a list of symbols
+typedef std::vector<symbol_t> symbol_list_t;
+
+// describe the reward accumulated by an agent
+typedef double reward_t;
+
+// describe a percept (observation or reward)
+typedef unsigned int percept_t;
+
+// describe the lifetime of an agent
+typedef unsigned long long lifetime_t;
+
+// describes an agent action
+typedef unsigned int action_t;
+
+// the program's keyword/value option pairs
+typedef std::map<std::string, std::string> options_t;
