+++ common/util.cpp
