+++ conf/coinflip.conf
